# docs/generated/openapi-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QAgent API Gateway",
    "description": "\n## QAgent API Gateway\n\nControl-plane API for QAgent — a native Agent Runtime for long-running Agent Teams\n(plan → execute → recover → deliver) with sandboxed tools and observability.\n\n### Features\n\n- **Models Management**: Query and retrieve available AI models\n- **MCP Configuration**: Manage Model Context Protocol (MCP) server configurations\n- **Memory Management**: Access and manage global memory data for personalized conversations\n- **Skills Management**: Query and manage skills and their enabled status\n- **Artifacts**: Access thread artifacts and generated files\n- **Health Monitoring**: System health check endpoints\n\n### Architecture\n\nLangGraph is mounted in-process under `/api/langgraph` (single uvicorn process).\nThis gateway provides custom endpoints for models, MCP configuration, skills, artifacts, and more.\n        ",
    "version": "1.0.0"
  },
  "paths": {
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List All Models",
        "description": "Retrieve a list of all available AI models configured in the system.",
        "operationId": "list_models_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Create Model",
        "description": "Add a new AI model to the configuration.",
        "operationId": "create_model_api_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/{model_name}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model Details",
        "description": "Retrieve detailed information about a specific AI model by its name.",
        "operationId": "get_model_api_models__model_name__get",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "models"
        ],
        "summary": "Update Model",
        "description": "Update an existing AI model configuration.",
        "operationId": "update_model_api_models__model_name__put",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Model",
        "description": "Remove an AI model from the configuration.",
        "operationId": "delete_model_api_models__model_name__delete",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Model Api Models  Model Name  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Unique identifier for the model"
          },
          "vendor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor",
            "description": "English vendor id (e.g. aliyun); one connection per vendor in YAML"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Actual provider model identifier"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Human-readable name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Model description"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "API key for the model"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Base URL for the API"
          },
          "use": {
            "type": "string",
            "title": "Use",
            "description": "Provider class to use",
            "default": "langchain_openai:ChatOpenAI"
          },
          "request_timeout": {
            "type": "number",
            "title": "Request Timeout",
            "description": "Request timeout in seconds",
            "default": 600.0
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries",
            "description": "Maximum number of retries",
            "default": 2
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "description": "Maximum number of tokens",
            "default": 8192
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "description": "Temperature for generation",
            "default": 0.7
          },
          "supports_vision": {
            "type": "boolean",
            "title": "Supports Vision",
            "description": "Whether model supports vision",
            "default": false
          },
          "supports_thinking": {
            "type": "boolean",
            "title": "Supports Thinking",
            "description": "Whether model supports thinking mode",
            "default": false
          },
          "supports_reasoning_effort": {
            "type": "boolean",
            "title": "Supports Reasoning Effort",
            "description": "Whether model supports reasoning effort",
            "default": false
          },
          "when_thinking_enabled": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "When Thinking Enabled",
            "description": "Extra ChatOpenAI kwargs when thinking is on (e.g. extra_body.thinking for OpenAI-compatible gateways)"
          },
          "thinking": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking",
            "description": "Shortcut merged into when_thinking_enabled.thinking (see config.example.yaml)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "model"
        ],
        "title": "ModelCreateRequest",
        "description": "Request model for creating/updating a model."
      },
...
      "ModelsListResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelResponse"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelsListResponse",
        "description": "Response model for listing all models."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}